requirements.txt
absl-py==0.15.0
astunparse==1.6.3
attrs==21.4.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
chumpy==0.70
ConfigArgParse==1.5.3
cycler==0.11.0
decorator==4.4.2
dominate==2.7.0
flatbuffers==1.12
fonttools==4.34.4
gast==0.3.3
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
imageio==2.19.5
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
joblib==1.1.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
llvmlite==0.38.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mediapipe==0.8.10.1
moviepy==1.0.3
numba==0.55.2
numpy==1.19.5
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
proglog==0.1.10
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-speech-features==0.6
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.15.0
sklearn==0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==3.1.0
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
typing-extensions==3.7.4.3
urllib3==1.26.10
Werkzeug==2.2.0
wrapt==1.12.1
zipp==3.8.1
